[dashboard] unskip platform/plugins/private/links/public/components/editor tests (#262904)
Closes https://github.com/elastic/kibana/issues/253303

Test skipped because it some times timed out (took over 5000ms). The
test is slow because it renders a lazy loaded component. This PR
resolves the lazy load by moving `DashboardNavigationOptionsEditor` into
a package. Unfortunately, this required creating 3 packages, `schema`,
`common`, and a `components` package.

The tests were originally added in
https://github.com/elastic/kibana/pull/211041.
https://github.com/elastic/kibana/pull/211041 fixed a bug but did not
resolve the code smell that caused the bug. The problem is that
`options` state is getting stored at 2 different levels in 2 different
`useStates`. This PR resolves that code smell by removing LinkOption
component and moving it into `LinkEditor` and then removing duplicated
`linkOptions` state.

Original test run (with lazy load component). Notice how first test case
takes over a second because lazy load component has to load.
<img width="1371" height="193" alt="Screenshot 2026-04-13 at 3 32 25 PM"
src="https://github.com/user-attachments/assets/0a9a58f6-b8ee-488d-ad9b-116467eb14a1"
/>

This PR test run (with component in package - so no lazy loading
needed).
<img width="1293" height="188" alt="Screenshot 2026-04-13 at 3 34 26 PM"
src="https://github.com/user-attachments/assets/5ff05378-ebe7-461f-a1e2-ecae57818944"
/>

---------

(cherry picked from commit e7743267aa62989fff197c20bbbaeb778d330a6f)
