+++ package.json
@@ -538,6 +538,9 @@
     "@kbn/dashboard-agent-common": "link:x-pack/platform/packages/shared/dashboard-agent/dashboard-agent-common",
     "@kbn/dashboard-agent-plugin": "link:x-pack/platform/plugins/shared/dashboard_agent",
     "@kbn/dashboard-markdown": "link:src/platform/plugins/shared/dashboard_markdown",
+    "@kbn/dashboard-navigation-options-common": "link:src/platform/packages/private/dashboard/dashboard-navigation-options-common",
+    "@kbn/dashboard-navigation-options-components": "link:src/platform/packages/private/dashboard/dashboard-navigation-options-components",
+    "@kbn/dashboard-navigation-options-schema": "link:src/platform/packages/private/dashboard/dashboard-navigation-options-schema",
     "@kbn/dashboard-plugin": "link:src/platform/plugins/shared/dashboard",
     "@kbn/dashboards-selector": "link:src/platform/packages/shared/dashboards/dashboards-selector",
     "@kbn/data-forge": "link:x-pack/platform/packages/shared/kbn-data-forge",
