+++ src/platform/packages/private/dashboard/dashboard-navigation-options-common/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/dashboard-navigation-options-common",
+  "owner": "@elastic/kibana-presentation",
+  "group": "platform",
+  "visibility": "private"
+}
