+++ src/platform/packages/private/dashboard/dashboard-navigation-options-common/moon.yml
@@ -1,0 +1,33 @@
+# This file is generated by the @kbn/moon package. Any manual edits will be erased!
+#  To extend this, write your extensions/overrides to 'moon.extend.yml'
+#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/dashboard-navigation-options-common'
+
+$schema: https://moonrepo.dev/schemas/project.json
+id: '@kbn/dashboard-navigation-options-common'
+layer: unknown
+owners:
+  defaultOwner: '@elastic/kibana-presentation'
+toolchains:
+  default: node
+language: typescript
+project:
+  title: '@kbn/dashboard-navigation-options-common'
+  description: Moon project for @kbn/dashboard-navigation-options-common
+  channel: ''
+  owner: '@elastic/kibana-presentation'
+  sourceRoot: src/platform/packages/private/dashboard/dashboard-navigation-options-common
+dependsOn: []
+tags:
+  - shared-common
+  - package
+  - prod
+  - group-platform
+  - private
+  - jest-unit-tests
+fileGroups:
+  src:
+    - '**/*.ts'
+    - '!target/**/*'
+  jest-config:
+    - jest.config.js
+tasks: {}
