+++ src/platform/packages/private/dashboard/dashboard-navigation-options-components/index.tsx
@@ -10,7 +10,7 @@
 import { i18n } from '@kbn/i18n';
 import React from 'react';
 import { EuiFormRow, EuiSpacer, EuiSwitch } from '@elastic/eui';
-import type { DashboardNavigationOptions } from '../../server';
+import type { DashboardNavigationOptions } from '@kbn/dashboard-navigation-options-schema';
 
 export interface Props {
   options: DashboardNavigationOptions;
