+++ src/platform/packages/private/dashboard/dashboard-navigation-options-components/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-browser",
+  "id": "@kbn/dashboard-navigation-options-components",
+  "owner": "@elastic/kibana-presentation",
+  "group": "platform",
+  "visibility": "private"
+}
