+++ src/platform/packages/private/dashboard/dashboard-navigation-options-schema/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-server",
+  "id": "@kbn/dashboard-navigation-options-schema",
+  "owner": "@elastic/kibana-presentation",
+  "group": "platform",
+  "visibility": "private"
+}
