+++ src/platform/packages/private/dashboard/dashboard-navigation-options-schema/moon.yml
@@ -1,0 +1,35 @@
+# This file is generated by the @kbn/moon package. Any manual edits will be erased!
+#  To extend this, write your extensions/overrides to 'moon.extend.yml'
+#  then regenerate this file with: 'node scripts/regenerate_moon_projects.js --update --filter @kbn/dashboard-navigation-options-schema'
+
+$schema: https://moonrepo.dev/schemas/project.json
+id: '@kbn/dashboard-navigation-options-schema'
+layer: unknown
+owners:
+  defaultOwner: '@elastic/kibana-presentation'
+toolchains:
+  default: node
+language: typescript
+project:
+  title: '@kbn/dashboard-navigation-options-schema'
+  description: Moon project for @kbn/dashboard-navigation-options-schema
+  channel: ''
+  owner: '@elastic/kibana-presentation'
+  sourceRoot: src/platform/packages/private/dashboard/dashboard-navigation-options-schema
+dependsOn:
+  - '@kbn/config-schema'
+  - '@kbn/dashboard-navigation-options-common'
+tags:
+  - shared-server
+  - package
+  - prod
+  - group-platform
+  - private
+  - jest-unit-tests
+fileGroups:
+  src:
+    - '**/*.ts'
+    - '!target/**/*'
+  jest-config:
+    - jest.config.js
+tasks: {}
