+++ src/platform/plugins/private/links/moon.yml
@@ -44,6 +44,9 @@ dependsOn:
   - '@kbn/es-query'
   - '@kbn/presentation-util'
   - '@kbn/kibana-utils-plugin'
+  - '@kbn/dashboard-navigation-options-common'
+  - '@kbn/dashboard-navigation-options-components'
+  - '@kbn/dashboard-navigation-options-schema'
 tags:
   - plugin
   - prod
