+++ src/platform/plugins/shared/dashboard/common/page_bundle_constants.ts
@@ -16,12 +16,6 @@ export const SEARCH_SESSION_ID = 'searchSessionId';
 /** The number of columns in the dashboard grid layout. */
 export const DASHBOARD_GRID_COLUMN_COUNT = 48;
 
-export const DEFAULT_DASHBOARD_NAVIGATION_OPTIONS = {
-  open_in_new_tab: false,
-  use_time_range: true,
-  use_filters: true,
-};
-
 // Do not change constant value - part of dashboard REST API
 export const DASHBOARD_DRILLDOWN_TYPE = 'dashboard_drilldown';
