+++ src/platform/plugins/shared/dashboard/public/dashboard_drilldown/dashboard_drilldown.ts
@@ -12,14 +12,12 @@ import { setStateToKbnUrl } from '@kbn/kibana-utils-plugin/public';
 import type { ApplyGlobalFilterActionContext } from '@kbn/unified-search-plugin/public';
 import { isFilterPinned } from '@kbn/es-query';
 import type { DrilldownDefinition } from '@kbn/embeddable-plugin/public/drilldowns/types';
+import { DEFAULT_DASHBOARD_NAVIGATION_OPTIONS } from '@kbn/dashboard-navigation-options-common';
 import type { DashboardDrilldownState } from '../../server/dashboard_drilldown/types';
 import { coreServices } from '../services/kibana_services';
 import { getLocation } from './get_location';
 import { cleanEmptyKeys } from '../../common/locator/locator';
-import {
-  DASHBOARD_DRILLDOWN_SUPPORTED_TRIGGERS,
-  DEFAULT_DASHBOARD_NAVIGATION_OPTIONS,
-} from '../../common/page_bundle_constants';
+import { DASHBOARD_DRILLDOWN_SUPPORTED_TRIGGERS } from '../../common/page_bundle_constants';
 import { DashboardDrilldownEditor } from './editor';
 
 export const dashboardDrilldown: DrilldownDefinition<
