+++ src/platform/plugins/shared/dashboard/public/dashboard_drilldown/editor.tsx
@@ -14,10 +14,10 @@ import { EuiFormRow, EuiComboBox, EuiSkeletonText } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import type { DrilldownEditorProps } from '@kbn/embeddable-plugin/public';
 import useDebounce from 'react-use/lib/useDebounce';
+import { DEFAULT_DASHBOARD_NAVIGATION_OPTIONS } from '@kbn/dashboard-navigation-options-common';
+import { DashboardNavigationOptionsEditor } from '@kbn/dashboard-navigation-options-components';
 import type { DashboardDrilldownState } from '../../server';
 import { findService } from '../dashboard_client';
-import { DEFAULT_DASHBOARD_NAVIGATION_OPTIONS } from '../../common/page_bundle_constants';
-import { DashboardNavigationOptionsEditor } from '../dashboard_navigation/options_editor';
 
 export const DashboardDrilldownEditor = (props: DrilldownEditorProps<DashboardDrilldownState>) => {
   const [options, setOptions] = useState<Array<EuiComboBoxOptionOption<string>>>([]);
