+++ src/platform/plugins/shared/dashboard/server/index.ts
@@ -61,9 +61,5 @@ export type {
 export type { DashboardDrilldownState } from './dashboard_drilldown/types';
 export type { DashboardSavedObjectAttributes, SavedDashboardPanel } from './dashboard_saved_object';
 export type { ScanDashboardsResult } from './scan_dashboards';
-export {
-  dashboardNavigationOptionsSchema,
-  type DashboardNavigationOptions,
-} from './dashboard_navigation';
 
 export { DASHBOARD_API_PATH } from '../common/constants';
