+++ src/platform/plugins/shared/dashboard/tsconfig.json
@@ -123,7 +123,10 @@
     "@kbn/core-http-common",
     "@kbn/core-http-server-mocks",
     "@kbn/core-http-browser",
-    "@kbn/lens-common"
+    "@kbn/lens-common",
+    "@kbn/dashboard-navigation-options-common",
+    "@kbn/dashboard-navigation-options-components",
+    "@kbn/dashboard-navigation-options-schema"
   ],
   "exclude": ["target/**/*"]
 }
