+++ tsconfig.base.json
@@ -906,6 +906,12 @@
       "@kbn/dashboard-agent-plugin/*": ["x-pack/platform/plugins/shared/dashboard_agent/*"],
       "@kbn/dashboard-markdown": ["src/platform/plugins/shared/dashboard_markdown"],
       "@kbn/dashboard-markdown/*": ["src/platform/plugins/shared/dashboard_markdown/*"],
+      "@kbn/dashboard-navigation-options-common": ["src/platform/packages/private/dashboard/dashboard-navigation-options-common"],
+      "@kbn/dashboard-navigation-options-common/*": ["src/platform/packages/private/dashboard/dashboard-navigation-options-common/*"],
+      "@kbn/dashboard-navigation-options-components": ["src/platform/packages/private/dashboard/dashboard-navigation-options-components"],
+      "@kbn/dashboard-navigation-options-components/*": ["src/platform/packages/private/dashboard/dashboard-navigation-options-components/*"],
+      "@kbn/dashboard-navigation-options-schema": ["src/platform/packages/private/dashboard/dashboard-navigation-options-schema"],
+      "@kbn/dashboard-navigation-options-schema/*": ["src/platform/packages/private/dashboard/dashboard-navigation-options-schema/*"],
       "@kbn/dashboard-plugin": ["src/platform/plugins/shared/dashboard"],
       "@kbn/dashboard-plugin/*": ["src/platform/plugins/shared/dashboard/*"],
       "@kbn/dashboards-selector": ["src/platform/packages/shared/dashboards/dashboards-selector"],
